testsys: refactor and allow aws region override in 'migration_crd'
Migration testing for VMware variants require a region for SSM to
dispatch upgrade/downgrade RunCommands to managed instances in vSphere.

Unlike ECS or EKS, TestSys cannot assume the cluster resource CRD
includes a region for vmware-k8s testing. Therefore we pass 'us-west-2'
as the override region when creating the migration test CRD.
